Unified-Cancer-Detection/requirements.txt
torch
torchvision
opencv-python-headless
numpy
pillow
pandas
fastapi
uvicorn
jinja2
python-multipart
# Added for metrics/analytics
matplotlib
scikit-learn
